+++ src/judy/src/JudyCommon/JudyInsArray.c
@@ -849,7 +849,7 @@ BuildBranch2:   // come here directly for Level = levelsub = cJU_ROOTSTATE.
         Pjp               = Pjbu->jbu_jp;           // for convenience in loop.
         numJPs            = 0;                      // non-null in the BranchU.
         digitmask         = cJU_MASKATSTATE(levelsub);   // see above.
-        digitshincr       = 1UL << (cJU_BITSPERBYTE * (levelsub - 1));
+        digitshincr       = (uintptr_t)1 << (cJU_BITSPERBYTE * (levelsub - 1));
         retval            = TRUE;
 
 // Scan and populate JPs (subexpanses):
